Moved TestObjectStore2 back into TestObjectStore now that HiveMetaStore has moved.
